refac: datalist 이후 컴포넌트, utils rename 및 arrow func 사용
