+++ test-web/src/components/Stats/Charts/Corr/Sens_Heated_Corr.js
@@ -3,12 +3,7 @@ import ApexCharts from 'react-apexcharts';
 import { statisticSensoryHeated } from '../../../../API/statistic/statisticSensoryHeated';
 import calculateChartSeries from './calculateChartSeries';
 
-export default function Sense_Heated_Corr({
-  startDate,
-  endDate,
-  animalType,
-  grade,
-}) {
+const Sense_Heated_Corr = ({ startDate, endDate, animalType, grade }) => {
   const [chartData, setChartData] = useState({});
   const [prop, setProp] = useState([]);
 
@@ -74,13 +69,13 @@ export default function Sense_Heated_Corr({
       enabled: true,
       y: {
         title: {
-          formatter: function (value, { seriesIndex, dataPointIndex, w }) {
+          formatter: (value, { seriesIndex, dataPointIndex, w }) => {
             const xLabel = w.globals.labels[dataPointIndex];
             const yLabel = w.config.series[seriesIndex].name;
             return `${yLabel} - ${xLabel}:`;
           },
         },
-        formatter: function (value) {
+        formatter: (value) => {
           const decimalValue = (value / 100).toFixed(3);
           return `${decimalValue}`;
         },
@@ -170,4 +165,6 @@ export default function Sense_Heated_Corr({
       height={350}
     />
   );
-}
+};
+
+export default Sense_Heated_Corr;
