+++ test-web/src/components/Stats/Charts/Corr/Sens_Proc_Corr.js
@@ -3,12 +3,7 @@ import ApexCharts from 'react-apexcharts';
 import { statisticSensoryProcessed } from '../../../../API/statistic/statisticSensoryProcessed';
 import calculateChartSeries from './calculateChartSeries';
 
-export default function Sense_Proc_Corr({
-  startDate,
-  endDate,
-  animalType,
-  grade,
-}) {
+const Sense_Proc_Corr = ({ startDate, endDate, animalType, grade }) => {
   const [chartData, setChartData] = useState({});
   const [prop, setProp] = useState([]);
 
@@ -73,13 +68,13 @@ export default function Sense_Proc_Corr({
       enabled: true,
       y: {
         title: {
-          formatter: function (value, { seriesIndex, dataPointIndex, w }) {
+          formatter: (value, { seriesIndex, dataPointIndex, w }) => {
             const xLabel = w.globals.labels[dataPointIndex];
             const yLabel = w.config.series[seriesIndex].name;
             return `${yLabel} - ${xLabel}:`;
           },
         },
-        formatter: function (value) {
+        formatter: (value) => {
           const decimalValue = (value / 100).toFixed(3);
           return `${decimalValue}`;
         },
@@ -169,4 +164,6 @@ export default function Sense_Proc_Corr({
       height={350}
     />
   );
-}
+};
+
+export default Sense_Proc_Corr;
