+++ test-web/src/components/Stats/Charts/Corr/Taste_Proc_Corr.js
@@ -2,13 +2,7 @@ import React, { useEffect, useState } from 'react';
 import ApexCharts from 'react-apexcharts';
 import { statisticProbexptProcessed } from '../../../../API/statistic/statisticProbexptProcessed';
 import calculateChartSeries from './calculateChartSeries';
-
-export default function Taste_Proc_Corr({
-  startDate,
-  endDate,
-  animalType,
-  grade,
-}) {
+const Taste_Proc_Corr = ({ startDate, endDate, animalType, grade }) => {
   const [chartData, setChartData] = useState({});
   const [prop, setProp] = useState([]);
 
@@ -73,13 +67,13 @@ export default function Taste_Proc_Corr({
       enabled: true,
       y: {
         title: {
-          formatter: function (value, { seriesIndex, dataPointIndex, w }) {
+          formatter: (value, { seriesIndex, dataPointIndex, w }) => {
             const xLabel = w.globals.labels[dataPointIndex];
             const yLabel = w.config.series[seriesIndex].name;
             return `${yLabel} - ${xLabel}:`;
           },
         },
-        formatter: function (value) {
+        formatter: (value) => {
           const decimalValue = (value / 100).toFixed(3);
           return `${decimalValue}`;
         },
@@ -169,4 +163,6 @@ export default function Taste_Proc_Corr({
       height={350}
     />
   );
-}
+};
+
+export default Taste_Proc_Corr;
