+++ test-web/src/components/Stats/Charts/HeatMap/Sens_Fresh_Map.js
@@ -3,12 +3,7 @@ import React, { useEffect, useState } from 'react';
 import { statisticSensoryFresh } from '../../../../API/statistic/statisticSensoryFresh';
 import calculateHeatMapChartSeries from './calculateHeatMapChartSeries';
 
-export default function Sens_Fresh_Map({
-  startDate,
-  endDate,
-  animalType,
-  grade,
-}) {
+const Sens_Fresh_Map = ({ startDate, endDate, animalType, grade }) => {
   const [chartData, setChartData] = useState({});
   const [prop, setProp] = useState([]);
 
@@ -83,7 +78,7 @@ export default function Sens_Fresh_Map({
       enabled: true,
       y: {
         title: {
-          formatter: function (value, { seriesIndex }) {
+          formatter: (value, { seriesIndex }) => {
             const axisName = ChartSeries[seriesIndex].name;
             const originalProperty = Object.keys(y_axis).find(
               (key) => y_axis[key] === axisName
@@ -91,7 +86,7 @@ export default function Sens_Fresh_Map({
             return `${axisName}(${originalProperty}):`;
           },
         },
-        formatter: function (value, { seriesIndex, dataPointIndex }) {
+        formatter: (value, { seriesIndex, dataPointIndex }) => {
           const count = ChartSeries[seriesIndex].data[dataPointIndex] || 0;
           const total =
             ChartSeries[seriesIndex].data.reduce((a, b) => a + b, 0) || 1;
@@ -101,7 +96,7 @@ export default function Sens_Fresh_Map({
       },
       x: {
         show: true,
-        formatter: function (value, { dataPointIndex, seriesIndex }) {
+        formatter: (value, { dataPointIndex, seriesIndex }) => {
           const categories = [
             '1 미만',
             '1.0 ~ 2.0',
@@ -131,4 +126,6 @@ export default function Sens_Fresh_Map({
       height={350}
     />
   );
-}
+};
+
+export default Sens_Fresh_Map;
