+++ test-web/src/components/Stats/Charts/HeatMap/Sens_Heated_Map.js
@@ -3,12 +3,7 @@ import React, { useEffect, useState } from 'react';
 import { statisticSensoryHeated } from '../../../../API/statistic/statisticSensoryHeated';
 import calculateHeatMapChartSeries from './calculateHeatMapChartSeries';
 
-export default function Sens_Heated_Map({
-  startDate,
-  endDate,
-  animalType,
-  grade,
-}) {
+const Sens_Heated_Map = ({ startDate, endDate, animalType, grade }) => {
   const [chartData, setChartData] = useState({});
   const [prop, setProp] = useState([]);
 
@@ -26,7 +21,6 @@ export default function Sens_Heated_Map({
       const data = await response.json();
       setProp(Object.keys(data));
       setChartData(data);
-      console.log(data);
     } catch (error) {
       console.error('Error fetching data:', error);
     }
@@ -83,7 +77,7 @@ export default function Sens_Heated_Map({
       enabled: true,
       y: {
         title: {
-          formatter: function (value, { seriesIndex }) {
+          formatter: (value, { seriesIndex }) => {
             const axisName = ChartSeries[seriesIndex].name;
             const originalProperty = Object.keys(y_axis).find(
               (key) => y_axis[key] === axisName
@@ -91,7 +85,7 @@ export default function Sens_Heated_Map({
             return `${axisName}(${originalProperty}):`;
           },
         },
-        formatter: function (value, { seriesIndex, dataPointIndex }) {
+        formatter: (value, { seriesIndex, dataPointIndex }) => {
           const count = ChartSeries[seriesIndex].data[dataPointIndex] || 0;
           const total =
             ChartSeries[seriesIndex].data.reduce((a, b) => a + b, 0) || 1;
@@ -101,7 +95,7 @@ export default function Sens_Heated_Map({
       },
       x: {
         show: true,
-        formatter: function (value, { dataPointIndex, seriesIndex }) {
+        formatter: (value, { dataPointIndex, seriesIndex }) => {
           const categories = [
             '1 미만',
             '1.0 ~ 2.0',
@@ -131,4 +125,6 @@ export default function Sens_Heated_Map({
       height={350}
     />
   );
-}
+};
+
+export default Sens_Heated_Map;
