+++ test-web/src/components/Stats/Charts/HeatMap/Taste_Fresh_Map.js
@@ -3,12 +3,7 @@ import React, { useEffect, useState } from 'react';
 import { statisticProbexptFresh } from '../../../../API/statistic/statisticProbexptFresh';
 import calculateHeatMapChartSeries from './calculateHeatMapChartSeries';
 
-export default function Taste_Fresh_Map({
-  startDate,
-  endDate,
-  animalType,
-  grade,
-}) {
+const Taste_Fresh_Map = ({ startDate, endDate, animalType, grade }) => {
   const [chartData, setChartData] = useState({});
   const [prop, setProp] = useState([]);
 
@@ -81,7 +76,7 @@ export default function Taste_Fresh_Map({
       enabled: true,
       y: {
         title: {
-          formatter: function (value, { seriesIndex }) {
+          formatter: (value, { seriesIndex }) => {
             const axisName = ChartSeries[seriesIndex].name;
             const originalProperty = Object.keys(y_axis).find(
               (key) => y_axis[key] === axisName
@@ -89,7 +84,7 @@ export default function Taste_Fresh_Map({
             return `${axisName}(${originalProperty}):`;
           },
         },
-        formatter: function (value, { seriesIndex, dataPointIndex }) {
+        formatter: (value, { seriesIndex, dataPointIndex }) => {
           const count = ChartSeries[seriesIndex].data[dataPointIndex] || 0;
           const total =
             ChartSeries[seriesIndex].data.reduce((a, b) => a + b, 0) || 1;
@@ -99,7 +94,7 @@ export default function Taste_Fresh_Map({
       },
       x: {
         show: true,
-        formatter: function (value, { dataPointIndex, seriesIndex }) {
+        formatter: (value, { dataPointIndex, seriesIndex }) => {
           const categories = [
             '1 미만',
             '1.0 ~ 2.0',
@@ -129,4 +124,6 @@ export default function Taste_Fresh_Map({
       height={350}
     />
   );
-}
+};
+
+export default Taste_Fresh_Map;
