+++ test-web/src/components/User/UserContext.js
@@ -4,7 +4,7 @@ import { createContext, useContext, useState } from 'react';
 
 const UserContext = createContext();
 
-export function UserProvider({ children }) {
+const UserProvider = ({ children }) => {
   const [user, setUser] = useState({}); // Your initial user state here
 
   return (
@@ -12,12 +12,13 @@ export function UserProvider({ children }) {
       {children}
     </UserContext.Provider>
   );
-}
+};
 
-export function useUser() {
+const useUser = () => {
   const context = useContext(UserContext);
   if (!context) {
     throw new Error('useUser must be used within a UserProvider');
   }
   return context.user;
-}
+};
+export { UserProvider, useUser };
